# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MindMate AI Local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ai/provider/key": {
      "post": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Save Ai Provider Key",
        "operationId": "save_ai_provider_key_api_v1_ai_provider_key_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Delete Ai Provider Key",
        "operationId": "delete_ai_provider_key_api_v1_ai_provider_key_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/provider/test": {
      "post": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Test Ai Provider Connection",
        "operationId": "test_ai_provider_connection_api_v1_ai_provider_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/provider/generation-mode": {
      "post": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Set Ai Generation Mode",
        "operationId": "set_ai_generation_mode_api_v1_ai_provider_generation_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/consent": {
      "get": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Get External Ai Consent",
        "operationId": "get_external_ai_consent_api_v1_ai_consent_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Accept External Ai Consent",
        "operationId": "accept_external_ai_consent_api_v1_ai_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/provider/openai/key": {
      "post": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Save Openai Provider Key",
        "operationId": "save_openai_provider_key_api_v1_ai_provider_openai_key_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Delete Openai Provider Key",
        "operationId": "delete_openai_provider_key_api_v1_ai_provider_openai_key_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/provider/openai/test": {
      "post": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Test Openai Provider Connection",
        "operationId": "test_openai_provider_connection_api_v1_ai_provider_openai_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/provider/openai/consent": {
      "get": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Get Openai Consent",
        "operationId": "get_openai_consent_api_v1_ai_provider_openai_consent_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ai-provider"
        ],
        "summary": "Accept Openai Consent",
        "operationId": "accept_openai_consent_api_v1_ai_provider_openai_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiProviderStatusResponse": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "credential_store": {
            "$ref": "#/components/schemas/CredentialStoreResponse"
          },
          "requested_model": {
            "type": "string",
            "title": "Requested Model"
          },
          "consent": {
            "$ref": "#/components/schemas/ConsentResponse"
          },
          "probe": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectionProbeResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "pricing_url": {
            "type": "string",
            "title": "Pricing Url"
          },
          "generation_mode": {
            "type": "string",
            "enum": [
              "mock",
              "deepseek",
              "openai_gpt6_sol"
            ],
            "title": "Generation Mode",
            "default": "mock"
          },
          "cost_estimate": {
            "$ref": "#/components/schemas/CostEstimateResponse"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/OnlineProviderCard"
            },
            "type": "array",
            "title": "Providers"
          },
          "learning_notice": {
            "type": "string",
            "title": "Learning Notice",
            "default": ""
          },
          "account_notice": {
            "type": "string",
            "title": "Account Notice",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "provider",
          "display_name",
          "configured",
          "credential_store",
          "requested_model",
          "consent",
          "source_url",
          "pricing_url",
          "cost_estimate"
        ],
        "title": "AiProviderStatusResponse"
      },
...
      "ApiKeyRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "format": "password",
            "title": "Api Key",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "ApiKeyRequest"
      },
...
      "ConnectionProbeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "failed"
            ],
            "title": "Status"
          },
          "checked_at": {
            "type": "string",
            "title": "Checked At"
          },
          "requested_model": {
            "type": "string",
            "title": "Requested Model"
          },
          "resolved_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Model"
          },
          "stream_supported": {
            "type": "string",
            "enum": [
              "supported",
              "unsupported",
              "unknown"
            ],
            "title": "Stream Supported"
          },
          "usage_supported": {
            "type": "string",
            "enum": [
              "supported",
              "unsupported",
              "unknown"
            ],
            "title": "Usage Supported"
          },
          "structured_output_supported": {
            "type": "string",
            "enum": [
              "supported",
              "unsupported",
              "unknown"
            ],
            "title": "Structured Output Supported"
          },
          "provider_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Request Id"
          },
          "response_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Fingerprint"
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          }
        },
        "type": "object",
        "required": [
          "status",
          "checked_at",
          "requested_model",
          "stream_supported",
          "usage_supported",
          "structured_output_supported"
        ],
        "title": "ConnectionProbeResponse"
      },
      "ConnectionTestRequest": {
        "properties": {
          "confirm_external_transfer": {
            "type": "boolean",
            "title": "Confirm External Transfer",
            "default": false
          }
        },
        "type": "object",
        "title": "ConnectionTestRequest"
      },
      "ConsentRequest": {
        "properties": {
          "version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "ConsentRequest"
      },
      "ConsentResponse": {
        "properties": {
          "current_version": {
            "type": "string",
            "title": "Current Version"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          }
        },
        "type": "object",
        "required": [
          "current_version",
          "accepted"
        ],
        "title": "ConsentResponse"
      },
...
      "CostEstimateResponse": {
        "properties": {
          "checked_on": {
            "type": "string",
            "title": "Checked On"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "pricing_url": {
            "type": "string",
            "title": "Pricing Url"
          },
          "rate_assumption": {
            "type": "string",
            "title": "Rate Assumption"
          },
          "input_usd_per_million_tokens": {
            "type": "string",
            "title": "Input Usd Per Million Tokens"
          },
          "output_usd_per_million_tokens": {
            "type": "string",
            "title": "Output Usd Per Million Tokens"
          },
          "knowledge_input_token_cap": {
            "type": "integer",
            "title": "Knowledge Input Token Cap"
          },
          "knowledge_output_token_cap": {
            "type": "integer",
            "title": "Knowledge Output Token Cap"
          },
          "knowledge_question_estimated_usd_ceiling": {
            "type": "string",
            "title": "Knowledge Question Estimated Usd Ceiling"
          },
          "probe_output_token_cap": {
            "type": "integer",
            "title": "Probe Output Token Cap"
          },
          "probe_estimated_usd_ceiling": {
            "type": "string",
            "title": "Probe Estimated Usd Ceiling"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "checked_on",
          "model",
          "pricing_url",
          "rate_assumption",
          "input_usd_per_million_tokens",
          "output_usd_per_million_tokens",
          "knowledge_input_token_cap",
          "knowledge_output_token_cap",
          "knowledge_question_estimated_usd_ceiling",
          "probe_output_token_cap",
          "probe_estimated_usd_ceiling",
          "disclaimer"
        ],
        "title": "CostEstimateResponse"
      },
      "CredentialStoreResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "type": "object",
        "required": [
          "available",
          "type"
        ],
        "title": "CredentialStoreResponse"
      },
...
      "GenerationModeRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "deepseek",
              "openai_gpt6_sol"
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "GenerationModeRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}